engine: introduce scene/runtime/draw2d, retained UI rebuild, doc consolidation
- engine/render: handle-based Server (Mesh/Material/Light/InstanceGroup);
  RenderScene/EndFrame split for overlay injection. Drop forward_renderer/
  ui_renderer/program_ui/gltf — folded into Server + draw2d.
- engine/scene: retained NodeID slab tree; PrepareFrame propagates transforms;
  scene calls renderer verbs but doesn't own it.
- engine/runtime: Host owns backend+renderer+scene+pipeline+fonts+events;
  View orchestrates one frame, delegates Render to RenderPipeline so multi-
  view stays an option.
- engine/draw2d: Context (record) + Renderer (submit) for retained 2D.
- engine/camera: Orbit + screen-ray pulled out of game.
- engine/gameapi: thin ABI shim (Init/Frame/Cleanup/InputEvent/WindowEvent).
- engine/ui: ui.Root + Node tree; primitives (Window, Padding, Flex, Grid,
  Align, AnchorPanel, ScrollView, Stack, SizedBox, Label, TextInput,
  FPSCounter); float32 lp end-to-end; DPI-only responsive viewport.
- engine/text: Font.{MeasureLine,Line,VolatileLine}; FontSet path-deduped;
  per-platform maxCachedLines.
- game: split per-platform game_api_exports + asset_path; board_view +
  game_input + game_view_meshes carved out of game.go; HUD as plain ui
  primitives + game-side controllers.
- ai/: drop shipped plan docs; tighten design.md/text-input.md; CLAUDE.md
  reduced to hard-won invariants only.

+++ engine/text/atlas_native.go
@@ -154,13 +154,13 @@ func (a *glyphAtlas) shapeRun(h int32, s string, ascentPx int32) ([]lineQuad, em
 			penX += glyph.XAdvance26_6
 			continue
 		}
-		x0 := (penX+glyph.XOffset26_6)/64.0 + entry.bitmap.BearingXPx
-		y0 := ascentPx - entry.bitmap.BearingYPx - glyph.YOffset26_6/64.0
+		x0 := (penX+glyph.XOffset26_6)/64 + entry.bitmap.BearingXPx
+		y0 := ascentPx - entry.bitmap.BearingYPx - glyph.YOffset26_6/64
 		w := entry.bitmap.WidthPx
 		hgt := entry.bitmap.HeightPx
 
 		out = append(out, lineQuad{
-			Dst: emath.Rect{X: x0, Y: y0, W: w, H: hgt},
+			Dst: emath.Rect{X: float32(x0), Y: float32(y0), W: float32(w), H: float32(hgt)},
 			UV: emath.UVRect{
 				U0: float32(entry.x) / atlasF,
 				V0: float32(entry.y) / atlasF,
@@ -204,5 +204,5 @@ func normalizeLineQuads(glyphs []lineQuad) ([]lineQuad, emath.Vec2) {
 		g.Dst.Y -= minY
 		out[i] = g
 	}
-	return out, emath.Vec2{float32(maxX - minX), float32(maxY - minY)}
+	return out, emath.Vec2{maxX - minX, maxY - minY}
 }
